Seimas.v2/skaidrumas/requirements.txt
httpx==0.27.0
lxml==5.2.2
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
playwright==1.44.0
schedule==1.2.1
python-dotenv==1.0.1
loguru==0.7.2
neo4j==5.20.0
rich==13.7.1
deepdiff==7.0.1
rapidfuzz==3.9.0
numpy==1.26.4
scipy==1.14.0
scikit-learn==1.5.1
networkx==3.3
